Move all Controller-related files to their own dir.
